backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.1
openai-whisper==20231117
numpy==1.24.3
torch==2.1.1
ffmpeg-python==0.2.0
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.23
aiosqlite==0.19.0 